Mentioned bugs have been fixed
