+++ core/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
-from .models import (
-    PhonePrefix, WareHouse, Currency, Country, ContactUs
+from core.models import (
+    PhonePrefix, WareHouse, Currency, Country, ContactUs, News
     )
 
 admin.site.register(PhonePrefix)
@@ -8,4 +8,5 @@ admin.site.register(WareHouse)
 admin.site.register(Currency)
 admin.site.register(Country)
 admin.site.register(ContactUs)
+admin.site.register(News)
