+++ user/admin.py
@@ -1,6 +1,5 @@
 from django.contrib import admin
-from .models import User, News, Wallet
+from user.models import User, Wallet
 
 admin.site.register(User)
 admin.site.register(Wallet)
-admin.site.register(News)
